getValue callback for Map and Search consistent with AutoComplete
